+++ CHANGELOG.md
@@ -17,12 +17,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Mutagenesis while no key or score cols is defined. this would be useful for mutant table in pure txt.
 - Mutagenesis with implicit chain identifier.
 - Multiple button locking.
+- Fix file handle mode of mutable txt button 
 
 ### Changed
 - Fetch repo path by dirname in test Scripts 
 - Move `get_color` as a public function in `utils`
 - Use `stick_radius` to present co-evolved pairs
+- Use explicit mutant description in Mutagenesis from PSSM profile
 
+
+### Removed
+- Mutant table checkpoint saving function. Replaced with mutant table saving function.
 
 ## [1.0.0] - 2023-09-24